Fixed wrong storage classifier (was LL instead of UL).
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@2841 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/bfs/Volume.cpp
@@ -208,10 +208,10 @@ Volume::Sync()
 status_t
 Volume::ValidateBlockRun(block_run run)
 {
-	if (run.allocation_group < 0 || run.allocation_group > AllocationGroups()
-		|| run.start > (1LL << AllocationGroupShift())
+	if (run.allocation_group < 0 || run.allocation_group > (int32)AllocationGroups()
+		|| run.start > (1UL << AllocationGroupShift())
 		|| run.length == 0
-		|| (uint32)run.length + run.start > (1LL << AllocationGroupShift())) {
+		|| uint32(run.length + run.start) > (1UL << AllocationGroupShift())) {
 		Panic();
 		FATAL(("*** invalid run(%ld,%d,%d)\n", run.allocation_group, run.start, run.length));
 		return B_BAD_DATA;
